refactoring: AdjustAbleColumn, PollColumn 관련 컴포넌트를 component 폴더 아래로 이동
